+++ PAF/snippets/core/PAF_Get-RecentFiles.ps1
@@ -5,6 +5,9 @@ Get recently modified files in a folder and its subfolders.
 .DESCRIPTION
 This function lists the most recently modified files present in the specified folder and its subfolders.
 
+.LINK
+https://github.com/voytas75/PowershellFramework
+The GitHub repository for the PowerShell Awesome Framework.
 #>